+++ lib/repostory/data.dart
@@ -1,0 +1,61 @@
+import 'package:visistafri/models/mostVisited.models.dart';
+import 'package:visistafri/models/place.models.dart';
+
+List<MostVisted> mostVisitedCategories = [
+    MostVisted('Buildings', [
+      Place('Ubumwe Grand', "assets/images/b1.jpg", 'Kigali,Rwanda',
+          'Ubwemwe Grand, Kigali Rwanda'),
+      Place('Ubumwe Grand', "assets/images/b2.jpg", 'Kigali,Rwanda',
+          'Ubwemwe Grand, Kigali Rwanda'),
+      Place('Ubumwe Grand', "assets/images/b3.jpg", 'Kigali,Rwanda',
+          'Ubwemwe Grand, Kigali Rwanda'),
+      Place('Ubumwe Grand', "assets/images/b1.jpg", 'Kigali,Rwanda',
+          'Ubumwe Grand, Kigali Rwanda'),
+      Place('Ubumwe Grand', "assets/images/b2.jpg", 'Kigali,Rwanda',
+          'Ubwemwe Grand, Kigali Rwanda'),
+      Place('Ubumwe Grand', "assets/images/b3.jpg", 'Kigali,Rwanda',
+          'Ubwemwe Grand, Kigali Rwanda'),
+    ]),
+    MostVisted('Parks', [
+      Place('Nyugwe Park', "assets/images/park3.jpg", 'Kigali,Rwanda',
+          'Nyugwe National Park'),
+      Place('Nyugwe Park', "assets/images/park2.jpg", 'Kigali,Rwanda',
+          'Nyugwe Park'),
+      Place('Nyugwe Park', "assets/images/park3.jpg", 'Kigali,Rwanda',
+          'Nyugwe Park'),
+      Place('Nyugwe Park', "assets/images/park3.jpg", 'Kigali,Rwanda',
+          'Nyugwe Park'),
+      Place('Nyugwe Park', "assets/images/park2.jpg", 'Kigali,Rwanda',
+          'Nyugwe Park'),
+      Place('Nyugwe Park', "assets/images/park3.jpg", 'Kigali,Rwanda',
+          'Nyugwe National Park'),
+    ]),
+    MostVisted('Lakes', [
+      Place('Lake Kivu', "assets/images/l1.jpg", 'Kigali,Rwanda',
+          'Lake Kivu'),
+      Place('Lake Kivu', "assets/images/l2.jpg", 'Kigali,Rwanda',
+          'Lake Kivu'),
+      Place('Lake Kivu', "assets/images/l3.jpg", 'Kigali,Rwanda',
+          'Lake Kivu'),
+       Place('Lake Kivu', "assets/images/l1.jpg", 'Kigali,Rwanda',
+          'Lake Kivu'),
+      Place('Lake Kivu', "assets/images/l2.jpg", 'Kigali,Rwanda',
+          'Lake Kivu'),
+      Place('Lake Kivu', "assets/images/l3.jpg", 'Kigali,Rwanda',
+          'Lake Kivu'),
+    ]),
+    MostVisted('Museums', [
+      Place('King Palace', "assets/images/m3.png", 'Kigali,Rwanda',
+          'King Palace'),
+      Place('King Palace', "assets/images/m2.png", 'Kigali,Rwanda',
+          'Genocide Memorial'),
+      Place('King Palace', "assets/images/m3.png", 'Kigali,Rwanda',
+          'King Palace'),
+       Place('King Palace', "assets/images/m3.png", 'Kigali,Rwanda',
+          'King Palace'),
+      Place('King Palace', "assets/images/m2.png", 'Kigali,Rwanda',
+          'Genocide Memorial'),
+      Place('King Palace', "assets/images/m3.png", 'Kigali,Rwanda',
+          'King Palace'),
+    ]),
+  ];
