--- lib/views/models/mostVisited.models.dart
@@ -1,7 +1,0 @@
-class Place {
-  String name;
-  String image;
-  String location;
-  String description;
-  Place(this.name, this.image, this.location, this.description);
-}
