--- lib/views/models/place.models.dart
@@ -1,10 +1,0 @@
-import 'package:visistafri/views/models/mostVisited.models.dart';
-
-class MostVisted {
-  String name;
-  List<Place> places;
-  MostVisted(
-    this.name,
-    this.places,
-  );
-}
